undidR: Difference-in-Differences with Unpoolable Data

Implements difference-in-differences with unpoolable data.

Details

A framework for estimating difference-in-differences with unpoolable data, based on Karim, Webb, Austin, and Strumpf (2025). Supports common or staggered adoption, multiple groups, and the inclusion of covariates. Also computes p-values for the aggregate average treatment effect on the treated via the randomization inference procedure described in MacKinnon and Webb (2020).

Stage One - Initialize

create_init_csv - Create initial CSV files for setup.

create_diff_df - Creates the "diff matrix".

Stage Two - Silos

undid_stage_two - Calculates trends and differences.

Stage Three - Analysis

undid_stage_three - Computes ATTs and p-values.

See also

Author

Eric Jamieson